+++ pipelines/training/autorag/documents_rag_optimization_pipeline/tests/functional/.env.example
@@ -44,10 +44,10 @@ LLAMA_STACK_VECTOR_IO_PROVIDER_ID_MILVUS_REMOTE=
 # FUNC_TEST_GENERATION_MODELS=
 
 # Functional test: filter test configs by tags (comma-separated, e.g. smoke,milvus-lite)
-# RHOAI_TEST_CONFIG_TAGS=
+# FUNCTIONAL_TESTS_TAGS=
 
 # =============================================================================
-# Env varialbes to test notebooks (can be extraced from kubernetes secrets as well)
+# Env variables to test notebooks (can be extracted from kubernetes secrets as well)
 # Note: these variables need to have the same names as one would expect in production
 # =============================================================================
 
@@ -56,7 +56,7 @@ LLAMA_STACK_CLIENT_BASE_URL=
 LLAMA_STACK_CLIENT_API_KEY=
 
 
-# AWS env varaibles for the indexing notebooks execution
+# AWS env variables for the indexing notebooks execution
 AWS_ACCESS_KEY_ID=
 AWS_SECRET_ACCESS_KEY=
 AWS_S3_ENDPOINT=
